fix(Legion Go): Improve new PID support
