+++ src/input/source/hidraw.rs
@@ -3,9 +3,7 @@ pub mod dualsense;
 pub mod flydigi_vader_4_pro;
 pub mod fts3528;
 pub mod horipad_steam;
-pub mod lego_dinput_combined;
-pub mod lego_dinput_split;
-pub mod lego_fps_mode;
+pub mod lego_touchpad;
 pub mod lego_xinput;
 pub mod legos_config;
 pub mod legos_imu;
@@ -38,15 +36,15 @@ use crate::{
     input::{
         capability::Capability, composite_device::client::CompositeDeviceClient,
         info::DeviceInfoRef, output_capability::OutputCapability,
+        source::hidraw::lego_touchpad::LegionTouchpadController,
     },
     udev::device::UdevDevice,
 };
 
 use self::{
     dualsense::DualSenseController, fts3528::Fts3528Touchscreen,
-    lego_dinput_combined::LegionControllerDCombined, lego_dinput_split::LegionControllerDSplit,
-    lego_fps_mode::LegionControllerFPS, lego_xinput::LegionControllerX,
-    legos_xinput::LegionSXInputController, opineo::OrangePiNeoTouchpad, steam_deck::DeckController,
+    lego_xinput::LegionXInputController, legos_xinput::LegionSXInputController,
+    opineo::OrangePiNeoTouchpad, steam_deck::DeckController,
 };
 
 use super::{InputError, OutputError, SourceDeviceCompatible, SourceDriver, SourceDriverOptions};
@@ -57,9 +55,7 @@ enum DriverType {
     DualSense,
     Fts3528Touchscreen,
     HoripadSteam,
-    LegionGoDCombined,
-    LegionGoDSplit,
-    LegionGoFPS,
+    LegionGoTouchpad,
     LegionGoSConfig,
     LegionGoSImu,
     LegionGoSTouchpad,
@@ -82,14 +78,12 @@ pub enum HidRawDevice {
     DualSense(SourceDriver<DualSenseController>),
     Fts3528Touchscreen(SourceDriver<Fts3528Touchscreen>),
     HoripadSteam(SourceDriver<HoripadSteam>),
-    LegionGoDCombined(SourceDriver<LegionControllerDCombined>),
-    LegionGoDSplit(SourceDriver<LegionControllerDSplit>),
-    LegionGoFPS(SourceDriver<LegionControllerFPS>),
+    LegionGoTouchpad(SourceDriver<LegionTouchpadController>),
     LegionGoSConfig(SourceDriver<LegionSConfigController>),
     LegionGoSImu(SourceDriver<LegionSImuController>),
     LegionGoSTouchpad(SourceDriver<LegionSTouchpadController>),
     LegionGoSXInput(SourceDriver<LegionSXInputController>),
-    LegionGoXInput(SourceDriver<LegionControllerX>),
+    LegionGoXInput(SourceDriver<LegionXInputController>),
     MsiClaw(SourceDriver<MsiClaw>),
     OrangePiNeo(SourceDriver<OrangePiNeoTouchpad>),
     RogAlly(SourceDriver<RogAlly>),
@@ -106,9 +100,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.info_ref(),
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.info_ref(),
             HidRawDevice::HoripadSteam(source_driver) => source_driver.info_ref(),
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.info_ref(),
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.info_ref(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.info_ref(),
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.info_ref(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.info_ref(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.info_ref(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.info_ref(),
@@ -130,9 +122,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.get_id(),
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.get_id(),
             HidRawDevice::HoripadSteam(source_driver) => source_driver.get_id(),
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.get_id(),
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.get_id(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.get_id(),
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.get_id(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.get_id(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.get_id(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.get_id(),
@@ -154,9 +144,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.client(),
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.client(),
             HidRawDevice::HoripadSteam(source_driver) => source_driver.client(),
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.client(),
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.client(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.client(),
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.client(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.client(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.client(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.client(),
@@ -178,9 +166,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.run().await,
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.run().await,
             HidRawDevice::HoripadSteam(source_driver) => source_driver.run().await,
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.run().await,
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.run().await,
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.run().await,
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.run().await,
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.run().await,
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.run().await,
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.run().await,
@@ -202,9 +188,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.get_capabilities(),
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.get_capabilities(),
             HidRawDevice::HoripadSteam(source_driver) => source_driver.get_capabilities(),
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.get_capabilities(),
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.get_capabilities(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.get_capabilities(),
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.get_capabilities(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.get_capabilities(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.get_capabilities(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.get_capabilities(),
@@ -228,11 +212,9 @@ impl SourceDeviceCompatible for HidRawDevice {
                 source_driver.get_output_capabilities()
             }
             HidRawDevice::HoripadSteam(source_driver) => source_driver.get_output_capabilities(),
-            HidRawDevice::LegionGoDCombined(source_driver) => {
+            HidRawDevice::LegionGoTouchpad(source_driver) => {
                 source_driver.get_output_capabilities()
             }
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.get_output_capabilities(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.get_output_capabilities(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.get_output_capabilities(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.get_output_capabilities(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => {
@@ -256,9 +238,7 @@ impl SourceDeviceCompatible for HidRawDevice {
             HidRawDevice::DualSense(source_driver) => source_driver.get_device_path(),
             HidRawDevice::Fts3528Touchscreen(source_driver) => source_driver.get_device_path(),
             HidRawDevice::HoripadSteam(source_driver) => source_driver.get_device_path(),
-            HidRawDevice::LegionGoDCombined(source_driver) => source_driver.get_device_path(),
-            HidRawDevice::LegionGoDSplit(source_driver) => source_driver.get_device_path(),
-            HidRawDevice::LegionGoFPS(source_driver) => source_driver.get_device_path(),
+            HidRawDevice::LegionGoTouchpad(source_driver) => source_driver.get_device_path(),
             HidRawDevice::LegionGoSConfig(source_driver) => source_driver.get_device_path(),
             HidRawDevice::LegionGoSImu(source_driver) => source_driver.get_device_path(),
             HidRawDevice::LegionGoSTouchpad(source_driver) => source_driver.get_device_path(),
@@ -334,26 +314,14 @@ impl HidRawDevice {
                 );
                 Ok(Self::SteamDeck(source_device))
             }
-            DriverType::LegionGoDCombined => {
-                let device = LegionControllerDCombined::new(device_info.clone())?;
+            DriverType::LegionGoTouchpad => {
+                let device = LegionTouchpadController::new(device_info.clone())?;
                 let source_device =
                     SourceDriver::new(composite_device, device, device_info.into(), conf);
-                Ok(Self::LegionGoDCombined(source_device))
-            }
-            DriverType::LegionGoDSplit => {
-                let device = LegionControllerDSplit::new(device_info.clone())?;
-                let source_device =
-                    SourceDriver::new(composite_device, device, device_info.into(), conf);
-                Ok(Self::LegionGoDSplit(source_device))
-            }
-            DriverType::LegionGoFPS => {
-                let device = LegionControllerFPS::new(device_info.clone())?;
-                let source_device =
-                    SourceDriver::new(composite_device, device, device_info.into(), conf);
-                Ok(Self::LegionGoFPS(source_device))
+                Ok(Self::LegionGoTouchpad(source_device))
             }
             DriverType::LegionGoXInput => {
-                let device = LegionControllerX::new(device_info.clone())?;
+                let device = LegionXInputController::new(device_info.clone())?;
                 let source_device =
                     SourceDriver::new(composite_device, device, device_info.into(), conf);
                 Ok(Self::LegionGoXInput(source_device))
@@ -518,32 +486,21 @@ impl HidRawDevice {
             return DriverType::SteamDeck;
         }
 
-        // Legion Go Dinput Combined
-        if vid == drivers::lego::driver_dinput_combined::VID
-            && drivers::lego::driver_dinput_combined::PIDS.contains(&pid)
+        // Legion Go Touchpad
+        if vid == drivers::lego::VID
+            && drivers::lego::PIDS.contains(&pid)
+            && iid == drivers::lego::TP_IID
         {
-            log::info!("Detected Legion Go DInput Combined Mode");
-            return DriverType::LegionGoDCombined;
-        }
-
-        // Legion Go Dinput Split
-        if vid == drivers::lego::driver_dinput_split::VID
-            && drivers::lego::driver_dinput_split::PIDS.contains(&pid)
-        {
-            log::info!("Detected Legion Go DInput Split Mode");
-            return DriverType::LegionGoDSplit;
-        }
-
-        // Legion Go FPS Mode
-        if vid == drivers::lego::driver_fps_mode::VID && drivers::lego::driver_fps_mode::PIDS.contains(&pid)
-        {
-            log::info!("Detected Legion Go FPS Mode");
-            return DriverType::LegionGoFPS;
+            log::info!("Detected Legion Go Touchpad");
+            return DriverType::LegionGoTouchpad;
         }
 
         // Legion Go XInput
-        if vid == drivers::lego::driver_xinput::VID && drivers::lego::driver_xinput::PIDS.contains(&pid) {
-            log::info!("Detected Legion Go XInput Mode");
+        if vid == drivers::lego::VID
+            && drivers::lego::PIDS.contains(&pid)
+            && iid == drivers::lego::GP_IID
+        {
+            log::info!("Detected Legion Go Controller");
             return DriverType::LegionGoXInput;
         }
 
@@ -651,5 +608,5 @@ impl HidRawDevice {
 
 /// Returns the DBus path for a [HIDRawDevice] from a device path (E.g. /dev/hidraw0)
 pub fn get_dbus_path(device_name: String) -> String {
-    format!("{}/{}", BUS_SOURCES_PREFIX, device_name)
+    format!("{BUS_SOURCES_PREFIX}/{device_name}")
 }
